Add Github Actions cron
This should run daily at 2:30 AM and run download, which as far as I can
tell, is a no-op if the files already exist. This does not yet commit
the repo if the results change.

This includes updates to main() to correctly pipe the credentials from
the default environment variable to the BigQuery client.